requirements.txt
# Core MCP and API
fastmcp>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# Data Processing
pydantic>=2.5.0
pyyaml>=6.0.1

# Caching (Optional)
redis>=5.0.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Security
sqlparse>=0.4.4
